pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soil-analytics"
version = "0.1.0"
description = "Soil analytics: FTIR/XRD/TGA reference checks and FESEM ML (Streamlit UI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.11",
  "pydantic>=2.5",
  "pyyaml>=6.0",
  "streamlit>=1.28",
  "plotly>=5.18",
  "scikit-learn>=1.3",
  "Pillow>=10.0",
  "torch>=2.1",
  "torchvision>=0.16",
  "timm>=0.9",
  "umap-learn>=0.5",
  "hdbscan>=0.8",
  "kagglehub>=0.2",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.1"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
